fixed circular dependency w/ TransactionCache & Cache class adding tests for adapters
