+++ packages/adapters/auctioneer/.nycrc.json
@@ -1,4 +1,3 @@
 {
-  "extends": "../../.nycrc.json",
   "exclude": ["src/index.ts", "test/**/*.ts"]
 }
