+++ packages/adapters/auctioneer/test/store.test.spec.ts
@@ -1,0 +1,10 @@
+
+import { expect } from "chai";
+import { TransactionsCache } from "@connext/nxtp-adapters-cache/src/lib/caches";
+import { Subscriptions } from "@connext/nxtp-adapters-cache/src/lib/entities";
+
+describe('Sequencer::Store Test', async () => {
+  const subscriptions: Subscriptions = new Map();
+  const transactions = new TransactionsCache({ url:"http://localhost:6379", subscriptions: subscriptions });
+  
+})
