+++ packages/agents/sequencer/src/index.ts
@@ -1,3 +1,6 @@
 import { makeSequencer } from "./sequencer";
 
+
 makeSequencer();
+
+export { makeSequencer };
